groundzero_app/landing_page_design_for_groundzero/workflows/requirements.txt
# Core dependencies
openai>=1.0.0
langgraph>=0.0.15
uagents>=0.9.1
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
langchain-core>=0.1.0
langchain>=0.1.0
graphviz>=0.20.1

# Server dependencies
fastapi>=0.109.0
uvicorn>=0.25.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0

# Utilities
colorama>=0.4.6
rich>=13.6.0